requirements.txt
asgiref==3.4.1
Django==3.2.11
django-filter==21.1
djangorestframework==3.13.1
gunicorn==20.1.0
importlib-metadata==4.10.0
Markdown==3.3.6
pkg_resources==0.0.0
pytz==2021.3
sqlparse==0.4.2
typing_extensions==4.0.1
whitenoise==5.3.0
zipp==3.7.0
